node添加用户名密码， service-gc-detail调整数据类型，添加定时任务采集数据，service添加分页关联查询

+++ src/main/java/com/easy/man/entity/po/ServiceGcDetail.java
@@ -1,5 +1,6 @@
 package com.easy.man.entity.po;
 
+import java.math.BigDecimal;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import java.time.LocalDateTime;
@@ -11,11 +12,11 @@ import lombok.experimental.Accessors;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author danny
- * @since 2019-01-07
+ * @since 2019-01-17
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -43,79 +44,79 @@ public class ServiceGcDetail implements Serializable {
     /**
      * 进程号
      */
-    private Integer pid;
+    private BigDecimal pid;
 
     /**
      * 年轻代中第一个survivor（幸存区）的容量 (字节)
      */
     @TableField("S0C")
-    private Integer s0c;
+    private BigDecimal s0c;
 
     /**
      * 年轻代中第二个survivor（幸存区）的容量 (字节)
      */
     @TableField("S1C")
-    private Integer s1c;
+    private BigDecimal s1c;
 
     /**
      * 年轻代中第一个survivor（幸存区）目前已使用空间 (字节)
      */
     @TableField("S0U")
-    private Integer s0u;
+    private BigDecimal s0u;
 
     /**
      * 年轻代中第二个survivor（幸存区）目前已使用空间 (字节)
      */
     @TableField("S1U")
-    private Integer s1u;
+    private BigDecimal s1u;
 
     /**
      * 年轻代中Eden（伊甸园）的容量 (字节)
      */
     @TableField("EC")
-    private Integer ec;
+    private BigDecimal ec;
 
     /**
      * 年轻代中Eden（伊甸园）目前已使用空间 (字节)
      */
     @TableField("EU")
-    private Integer eu;
+    private BigDecimal eu;
 
     /**
      * Old代的容量 (字节)
      */
     @TableField("OC")
-    private Integer oc;
+    private BigDecimal oc;
 
     /**
      * Old代目前已使用空间 (字节)
      */
     @TableField("OU")
-    private Integer ou;
+    private BigDecimal ou;
 
     /**
      * 方法区容量(字节)
      */
     @TableField("MC")
-    private Integer mc;
+    private BigDecimal mc;
 
     /**
      * 方法区已使用大小(字节)
      */
     @TableField("MU")
-    private Integer mu;
+    private BigDecimal mu;
 
     /**
      * 压缩类空间大小
      */
     @TableField("CCSC")
-    private Integer ccsc;
+    private BigDecimal ccsc;
 
     /**
      * 压缩类空间使用大小
      */
     @TableField("CCSU")
-    private Integer ccsu;
+    private BigDecimal ccsu;
 
     /**
      * 从应用程序启动到采样时年轻代中gc次数
@@ -127,7 +128,7 @@ public class ServiceGcDetail implements Serializable {
      * 从应用程序启动到采样时年轻代中gc所用时间(s)
      */
     @TableField("YGCT")
-    private Integer ygct;
+    private BigDecimal ygct;
 
     /**
      * 从应用程序启动到采样时old代(全gc)gc次数
@@ -139,13 +140,13 @@ public class ServiceGcDetail implements Serializable {
      * 从应用程序启动到采样时old代(全gc)gc所用时间(s)
      */
     @TableField("FGCT")
-    private Integer fgct;
+    private BigDecimal fgct;
 
     /**
      * 从应用程序启动到采样时gc用的总时间(s)
      */
     @TableField("GCT")
-    private Integer gct;
+    private BigDecimal gct;
 
     /**
      * 采样时间
